Implementation of Input System (#31)
Created the Input class, which is the abstracted class. This holds
virtual functions for pressed, released, and held for keyboard, mouse,
and gamepad, as well as tracking mouse position.

The GLFWInput class inherits from Input to hold the glfw-specific logic
to allow the inputs to actually work. Holds the states of the inputs
pressed, is able to check the states of the inputs, and updates the
states is pressed/held/released.

The InputSystem class that will be used as the access point for input
functionality. Uses an inputBackend pointer to access all of the
functionality from the Input/GLFWInput classes.