Script System: Init and Math Bindings (#85)
With these changes, Sol2/Lua are added to the engine, and I've written a
binding for our Math library which is in the scripting tests. Currently
it lacks intellisense, but I'm working on that. You can see example luas
and in Sol2API you can see how I've set up accessible libraries for our
Lua code.